Update Tools.java
When the gates, outputs, and powers are created they do not need to know what their inputs are.
There needs to be something like a hash table to store the name of the nodes as the key, and the actual node reference as the value.
This way the string names can easily be resolved to their node references.

After all the nodes are created, the xml needs iterated through again using the hash table to resolve the names into node references to properly connect the inputs.
Or when the nodes are created, a hash table with the name as the key and the linked list of input names as the value. Then the GateOr.NodeList can be iterated through and the name of the node can be used in that hash table to get the input names, then the input names can go through the hash table that stores the node references and uses those references as the input of the original node in the GateOr.NodeList.